wip:(5/n): refactor module ast
This is the last commit which will contain changes to the abstract
syntax tree, and is the one which answers the question: "what does a
parsed module look like?".

This commit introduces a new `Module` type, which supercedes the
previous `ProgramAst`, `ModuleAst` and `Module` types, providing all of
the information you might want to know about a given module, as well as
supporting the core functionality necessary to parse, serialize, and
pretty print modules.

A `Module` has a `ModuleKind`, which identifies what type of module it
represents: an executable (e.g. what was previously a `ProgramAst`), a
library (e.g. `what was previously a `ModuleAst`), or a kernel (which
previously had no specific representation).

The `ModuleKind` dictates the semantics of the module, but in practice
all the "kinds" of modules are virtually identical except for these
slight semantic differences, which is why they have been unified here.
These semantics are validated by the semantic analysis pass, and catch
the set of things you are not allowed to do in certain types of modules,
e.g. exporting from an executable module, syscall/call in a kernel
module, `begin` in a library module.

Lastly, this commit removes the old `ModuleImports` type, which is
superceded by these changes (as well as subsequent ones in other parts
of the assembler). In its place, is a new `Import` type which represents
all of the details about a specific module import. Each `Module` has a
set of `Import` associated with it, which it uses to resolve names, as
well as determine syntax-level inter-module dependencies.