fix: improve dynexec/procref test
The previous version of this test started to fail because it assumed
that it could blindly call u64::wrapping_add on a word representing the
hash of a procedure. This is not actually safe, as there is no guarantee
the hash will consist of 4 valid 32-bit limbs - and in fact that is what
caused it to start to fail here - the procedure hash changed, was no
longer valid for use as an operand to that function, and caused
execution to trap deep in the VM.

This commit changes the test to use actual checked operands, and adds a
separate module into the mix to facilitate testing that procref/dynexec
works across modules.